fix(tests): removed timeout, refactoring
