+++ tests/playwright.config.ts
@@ -1,0 +1,3 @@
+export const getPlayWrightGotoConfig = () => ({
+  timeout: 0,
+});
