--- tests/server/server.config.ts
@@ -1,1 +1,0 @@
-export const serverPort = 3073;
