+++ tests/vitejs.server.config.ts
@@ -1,0 +1,7 @@
+export const getViteServerConfig = (port = 3073) => ({
+  configFile: `${__dirname}/../vite.config.ts`,
+  root: `${__dirname}/../`,
+  server: {
+    port,
+  },
+});
